Sets up a logger This sets up a logger with the desired log level and format, and creates both a file handler and a stream handler to write log messages to a file and print them to the console, respectively. It then replaces all the calls to print with calls to the logger.info function, which logs the messages at the INFO level.
To use the logger in your code, you can simply call the appropriate logging function on the logger object. For example:
logger.debug('This is a debug message.')
logger.info('This is an info message.')
logger.warning('This is a warning message.')
logger.error('This is an error message.')
logger.critical('This is a critical message.')

These log messages will be written to the main.log file as well as printed to the console, depending on the log level that you have set.

In the code you provided, the log level is set to INFO, so only log messages at the INFO, WARNING, ERROR, and CRITICAL levels will be logged. If you want to log messages at other levels, you can set the log level to a different value. For example:
logger.setLevel(logging.DEBUG)

logger.setLevel(logging.WARNING)

Signed-off-by: Sakshi Aggarwal <81718060+sakshiagrwal@users.noreply.github.com>